separate parts of fit that differ
